+++ R/graphs.R
@@ -39,15 +39,15 @@ path_graphs <- function(.analysis_data,
   analysis_data <- .analysis_data %>%
     dplyr::filter(.data[[case_colname]] %in% cases)
   
-  # Calculate energy, cost, and preferences paths
+  # Calculate energy, expenditure, and preferences paths
   e_paths <- analysis_data %>%
     energy_paths(indexed = indexed, graph_params = graph_params)
-  c_paths <- analysis_data %>% 
-    cost_paths(indexed = indexed, graph_params = graph_params)
+  exp_paths <- analysis_data %>% 
+    expenditure_paths(indexed = indexed, graph_params = graph_params)
   p_paths <- analysis_data %>% 
     prefs_paths(graph_params = graph_params)
   # Bundle all paths together
-  paths <- dplyr::bind_rows(e_paths, c_paths, p_paths) %>% 
+  paths <- dplyr::bind_rows(e_paths, exp_paths, p_paths) %>% 
     dplyr::filter(.data[[graph_df_colnames$graph_type_col]] %in% graph_types)
   
   # Extract points between rebound effects
@@ -56,7 +56,7 @@ path_graphs <- function(.analysis_data,
                    rebound_stages = rebound_stages, 
                    rebound_segments = rebound_segments, 
                    graph_df_colnames = graph_df_colnames)
-  c_points <- c_paths %>% 
+  exp_points <- exp_paths %>% 
     extract_points(graph_params = graph_params, 
                    rebound_stages = rebound_stages, 
                    rebound_segments = rebound_segments, 
@@ -67,21 +67,21 @@ path_graphs <- function(.analysis_data,
                    rebound_segments = rebound_segments, 
                    graph_df_colnames = graph_df_colnames)
   # Bundle all points together
-  points <- dplyr::bind_rows(e_points, c_points, p_points) %>% 
+  points <- dplyr::bind_rows(e_points, exp_points, p_points) %>% 
     dplyr::filter(.data[[graph_df_colnames$graph_type_col]] %in% graph_types)
 
-  # Calculate energy, cost, and preferences grids/guide lines
+  # Calculate energy, expenditure, and preferences grids/guide lines
   e_grid_data <- analysis_data %>% 
     iso_energy_lines(indexed = indexed, graph_params = graph_params)
-  c_grid_data <- analysis_data %>% 
-    iso_cost_lines(indexed = indexed, graph_params = graph_params)
+  exp_grid_data <- analysis_data %>% 
+    iso_expenditure_lines(indexed = indexed, graph_params = graph_params)
   p_grid_data <- analysis_data %>% 
     iso_budget_lines_prefs(graph_params = graph_params)
   # Decide which grids we want to keep.
   # I.e., we should not keep grids for graphs that we're not making.
   keep_grids <- intersect(graph_types, grid_types)
   # Bundle them together
-  grids <- dplyr::bind_rows(e_grid_data, c_grid_data, p_grid_data) %>% 
+  grids <- dplyr::bind_rows(e_grid_data, exp_grid_data, p_grid_data) %>% 
     dplyr::filter(.data[[graph_df_colnames$graph_type_col]] %in% keep_grids)
 
   # Calculate indifference curves for the preferences graph  
@@ -120,7 +120,7 @@ path_graphs <- function(.analysis_data,
         ggplot2::ylab(expression(dot(E)[indir] * " [MJ/year]"))
     }
   }
-  if (graph_types == ReboundTools::graph_types$cost) {
+  if (graph_types == ReboundTools::graph_types$expenditure) {
     if (indexed) {
       g <- g +
         # Horizontal axis label C_dot_dir/C_dot_dir_orig
@@ -341,7 +341,7 @@ rebound_graphs_helper <- function(.path_data,
 #' @export
 #'
 #' @examples
-#' # Sensitivity of total rebound (Re_tot) to productivity multiplier (k)
+#' # Sensitivity of total rebound (Re_tot) to macro multiplier (k)
 #' df <- load_eeu_data()
 #' sens_params <- list(Car = list(k = seq(0.5, 1.5, by = 0.5)), 
 #'                     Lamp = list(k = seq(0, 2, by = 1)))
@@ -357,7 +357,7 @@ rebound_graphs_helper <- function(.path_data,
 #'                linetype = ggplot2::element_blank())
 #'                
 #' # A more-complicated example that shows multi-variate sensitivity.
-#' # Values of the productivity parameter (k) is shown in rows of the lattice plot.
+#' # Values of the macro parameter (k) is shown in rows of the lattice plot.
 #' # Uncompensated price elasticity of energy service consumption (e_qs_ps_UC) 
 #' # is shown in columns of the lattice plot.
 #' # Total rebound (Re_tot) is given on the y-axis, and 
@@ -388,7 +388,7 @@ rebound_graphs_helper <- function(.path_data,
 #'                       Lamp = list(eta_engr_units_star = seq(70, 90, by = 5)))
 #' # Choose rebound terms to include in the graph and their order
 #' rebound_vars <- c("Re_dempl", "Re_emb", "Re_md", "Re_dsub", "Re_isub", 
-#'                   "Re_dinc", "Re_iinc", "Re_prod")
+#'                   "Re_dinc", "Re_iinc", "Re_macro")
 #'                   
 #' sensitivity_graphs(rebound_data = df, parameterization = sens_params_3,
 #'                    x_var = "eta_engr_units_tilde", 
@@ -402,7 +402,7 @@ rebound_graphs_helper <- function(.path_data,
 #'           Re_isub = ReboundTools::path_graph_params$isub_colour, 
 #'           Re_dinc = ReboundTools::path_graph_params$dinc_colour,
 #'           Re_iinc = ReboundTools::path_graph_params$iinc_colour,
-#'           Re_prod = ReboundTools::path_graph_params$prod_colour), 
+#'           Re_macro = ReboundTools::path_graph_params$macro_colour), 
 #'                                breaks = rebound_vars) +
 #'  ggplot2::scale_size_manual(values = 
 #'         c(Re_dempl = 0.2, 
@@ -412,7 +412,7 @@ rebound_graphs_helper <- function(.path_data,
 #'           Re_isub = ReboundTools::path_graph_params$isub_size, 
 #'           Re_dinc = ReboundTools::path_graph_params$dinc_size,
 #'           Re_iinc = ReboundTools::path_graph_params$iinc_size,
-#'           Re_prod = ReboundTools::path_graph_params$prod_size), 
+#'           Re_macro = ReboundTools::path_graph_params$macro_size), 
 #'                             breaks = rebound_vars) +
 #' ggplot2::scale_linetype_manual(values = 
 #'         c(Re_dempl = ReboundTools::path_graph_params$dempl_linetype, 
@@ -422,7 +422,7 @@ rebound_graphs_helper <- function(.path_data,
 #'           Re_isub = "11",
 #'           Re_dinc = ReboundTools::path_graph_params$dinc_linetype,
 #'           Re_iinc = "11",
-#'           Re_prod = ReboundTools::path_graph_params$prod_linetype), 
+#'           Re_macro = ReboundTools::path_graph_params$macro_linetype), 
 #'                                breaks = rebound_vars) +
 #' ggplot2::labs(x = expression(tilde(eta)*" [mpg (Car) or lm/W (Lamp)]"), 
 #'               y = "Re terms [-]", 
@@ -596,7 +596,7 @@ rebound_terms_graph <- function(.parametric_data = parametric_analysis(rebound_d
                                             Re_dinc = graph_params$dinc_colour,
                                             Re_iinc = graph_params$iinc_colour,
                                             Re_inc = graph_params$inc_colour,
-                                            Re_prod = graph_params$prod_colour,
+                                            Re_macro = graph_params$macro_colour,
                                             Re_dir = graph_params$dir_colour,
                                             Re_indir = graph_params$indir_colour,
                                             Re_tot = graph_params$tot_colour), 
@@ -613,7 +613,7 @@ rebound_terms_graph <- function(.parametric_data = parametric_analysis(rebound_d
                                           Re_dinc = graph_params$dinc_size,
                                           Re_iinc = graph_params$iinc_size,
                                           Re_inc = graph_params$inc_size,
-                                          Re_prod = graph_params$prod_size,
+                                          Re_macro = graph_params$macro_size,
                                           Re_dir = graph_params$dir_size,
                                           Re_indir = graph_params$indir_size,
                                           Re_tot = graph_params$tot_size),
@@ -630,7 +630,7 @@ rebound_terms_graph <- function(.parametric_data = parametric_analysis(rebound_d
                                               Re_dinc = graph_params$dinc_linetype,
                                               Re_iinc = graph_params$iinc_linetype,
                                               Re_inc = graph_params$inc_linetype,
-                                              Re_prod = graph_params$prod_linetype, 
+                                              Re_macro = graph_params$macro_linetype, 
                                               Re_dir = graph_params$dir_linetype,
                                               Re_indir = graph_params$indir_linetype,
                                               Re_tot = graph_params$tot_linetype),