+++ R/points.R
@@ -5,7 +5,7 @@
 #' Each available stage of the rebound process is represented in the data frame.
 #'
 #' @param .paths A data frame of rebound paths, 
-#'                      likely created by `energy_paths()`, `cost_paths()`, or `prefs_paths()`.
+#'                      likely created by `energy_paths()`, `expenditure_paths()`, or `prefs_paths()`.
 #' @param graph_params See `ReboundTools::graph_params`.
 #' @param rebound_stages See `ReboundTools::rebound_stages`.
 #' @param rebound_segments See `ReboundTools::rebound_segments`.
@@ -22,7 +22,7 @@
 #'   extract_points()
 #' load_eeu_data() %>% 
 #'   rebound_analysis() %>% 
-#'   cost_paths() %>% 
+#'   expenditure_paths() %>% 
 #'   extract_points()
 #' load_eeu_data() %>% 
 #'   rebound_analysis() %>% 
@@ -65,11 +65,11 @@ extract_points <- function(.paths,
       # and eliminate the line names column.
       "{graph_df_colnames$point_name_col}" := dplyr::case_when(
         # Set the point name.
-        .data[[graph_df_colnames$line_name]] == rebound_segments$dempl ~ rebound_stages$orig, 
-        .data[[graph_df_colnames$line_name]] == rebound_segments$isub  ~ rebound_stages$star, 
-        .data[[graph_df_colnames$line_name]] == rebound_segments$dinc  ~ rebound_stages$hat,
-        .data[[graph_df_colnames$line_name]] == rebound_segments$prod  ~ rebound_stages$bar, 
-        .data[[graph_df_colnames$line_name]] == ls                     ~ lp
+        .data[[graph_df_colnames$line_name]] == rebound_segments$dempl  ~ rebound_stages$orig, 
+        .data[[graph_df_colnames$line_name]] == rebound_segments$isub   ~ rebound_stages$star, 
+        .data[[graph_df_colnames$line_name]] == rebound_segments$dinc   ~ rebound_stages$hat,
+        .data[[graph_df_colnames$line_name]] == rebound_segments$macro  ~ rebound_stages$bar, 
+        .data[[graph_df_colnames$line_name]] == ls                      ~ lp
       ), 
       # Eliminate unneeded columns
       "{graph_df_colnames$line_name_col}" := NULL,